Promote the `Path` type to a full data type
A `Path` is now a data-type in its own right - currently with a single
constructor that takes a list of `PathElement`s. This is currently isomorphic to
`[Text]` as it was before, but this will allow us to have richer information for
`listOf`. In the next commit I will add a new data constructor for paths into
lists which will allow better pattern matching (and less string typing!).