Use a `MetaPath` with `Index` elements for lists of forms.
Introduces the `MetaPath` constructor for `Path`s, which indicates that a `Path`
does not correspond directly to user input, but rather something that is
derived *about* a user's input.

Also introduces the `Index` path element type, which represents an index into an
indexed path, where an indexed path is something where there are multiple paths
depending on the index. Lists of forms make indexed paths - where every
occurance of a form submission is a new index.